backend/requirements.txt
asgiref==3.8.1
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.3
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1 
gunicorn==23.0.0 
psycopg2-binary==2.9.10  
whitenoise==6.8.2  
python-dotenv==1.0.1  
uvicorn==0.34.0  
dj-database-url==2.3.0 
psycopg==3.2.3
psycopg2===2.9.10
requests==2.31.0
google-generativeaiclear

